correct building loss

+++ spells.py
@@ -5,5 +5,5 @@ class Fireball(Spell):
 	mana_cost = 50
 	@staticmethod
 	def cast(target):
-		target.num_population = round(target.num_population * 0.9)
+		target.num_population = int(round(target.num_population * 0.9))
